security(tls): use INT_MAX for portable session restore length validation
Replace LONG_MAX with INT_MAX in SocketTLS_session_restore() bounds check
to ensure portability across all platforms, particularly 64-bit Windows
(LLP64 model) where sizeof(long)=4 but sizeof(size_t)=8.

The d2i_SSL_SESSION() OpenSSL function takes a long parameter, so we need
to ensure the size_t length fits safely in that range. Using INT_MAX
provides a conservative, platform-independent upper bound that:
- Prevents integer overflow during the cast to long
- Works correctly on LLP64 (Windows x64), LP64 (Unix x64), and ILP32
- Still allows session data up to ~2GB (more than sufficient)

Add comprehensive test coverage for edge cases including zero-length
and oversized session data.

Fixes #2354

+++ src/tls/SocketTLS.c
@@ -1757,8 +1757,12 @@ SocketTLS_session_restore (Socket_T socket, const unsigned char *buffer,
   if (!ssl)
     return -1;
 
-  /* Validate length to prevent integer overflow in d2i_SSL_SESSION */
-  if (len == 0 || len > (size_t)LONG_MAX)
+  /* Validate length bounds for d2i_SSL_SESSION (takes long parameter).
+   * Use INT_MAX for portability - it's safe on all platforms including
+   * 64-bit Windows (LLP64) where sizeof(long)=4 but sizeof(size_t)=8.
+   * This conservative check prevents both overflow and platform-specific
+   * edge cases while still allowing session data up to ~2GB. */
+  if (len == 0 || len > INT_MAX)
     return 0; /* Invalid data */
 
   /* Deserialize session from DER format */
